Fix conflict and updated based on comments
